Don't fail if apparmor_parser is missing (#40)
Currently a missing apparmor_parser binary leads to a panic at startup
already since the property value gets evaluated when the propery is
registred.

Set the property to an empty string if apparmor_parser is missing. This
will allow to generate a sensible warning in Supervisor so the user can
act on it.