Modernize go dev environment (#162)
* Use bookworm as dev container

* Use %w (wrapping format verb) for errors

This allows to enable errorlint.

* Avoid deprecated io/ioutil

* Disable gosec linter

The cmdline.txt is a system configuration, those are typically readable
on a Linux system. The file is on a FAT partition anyways, so the file
permissions are mandated by mount options anyways.

* Use Go version 1.19 by default

* Enable bugs linters

Enable the preset bugs plus enable the default ones which are not part
of the bugs preset explicitly.