Remove WipeDevice method (#225)
WipeDevice was called by the haos-wipe.service when the OS was marked
for wipe. However, that was removed in [1] and replaced by shell script
present on the OS itself, so there should be no longer any user of this
service. Remove it from OS Agent code to have only a single way for the
device wipe.

[1] https://github.com/home-assistant/operating-system/pull/3916