Remove the option to revert to `overlay2` driver (#245)
With update to Docker v29, the Containerd snapshotter driver is now used
by default for fresh installs. This means that once the overlay2 data is
purged, removing the flag file no longer results in Docker using the
overlay2 driver, and it keeps using overlayfs. Since this option was
added just out of convenience and with Docker's long term plan it makes
little sense to keep overlay2 as an option, just remove it from OS
agent.